feat: propogate .unnumbered and .notoc on captioned elements
Disabling element numbering and/or addition to the ToC (for lists
of figures or tables).
For captioned figures, it works with Markdown, Randoc AST and Djot.
For captioned tables, it works with djot only, as the latter allows
attributes on any element (here the table), but regular Markdown
does not have that capability.