feat: Customizable strike, underline, insertion, deletion and mark
First, make all these inline elements from Djot or Markdown all
equivalent to a corresponding span with certain classes, so both
cases are implemented identically. We thus have a consistent
mapping in all cases.
Then, use styles, if available, to render them. So for instance,
Markdown ~~xxx~~ is strictly equivalent to [xxx]{.strike}, and
to [xxx]{custom-style=md-strikethrough} in resilient context if
that style exists, for complete customization.