fix: Propagate Djot attributes to the blockquote implementation
So the latter might implement a different behavior depending on them.
As of now, SILE's default blockquote (from SILE 0.15.0) and resilient's
style-aware blockquote don't support options, so this would not do anything
yet, but it paves the way for more customization or alternate
implementations.
N.B. It's also what we were already doing for the captioned quotes
(via the epigraph environment, which does support several options), so
it also fixes a small API discrepancy...